[FIX] stock_picking_report_valued: prevent singleton error in `res.currency` for warehouses with multiple delivery steps
This report adds new fields of `Monetary` type. When the report is printed, Odoo tries to access `currency_id`, which is related to the `sale_line`. This field is only set on the final delivery step.

However, when the warehouse is configured with two or three delivery steps, the initial pickings contain move lines that are not linked to any sale line, which raises the following error:
ValueError: Expected singleton: res.currency()

This commit ensures that the fields are rendered using the picking currency, thus preventing the error.